Update codebase to upstream revision 0ee5b604af6b842abc977bc5c57e5791a5b9ceb0.

+++ include/zephyr/kernel/thread.h
@@ -111,7 +111,7 @@ struct _thread_base {
 
 #ifdef CONFIG_SCHED_CPU_MASK
 	/* "May run on" bits for each CPU */
-	uint16_t cpu_mask;
+	uint32_t cpu_mask;
 #endif /* CONFIG_SCHED_CPU_MASK */
 
 	/* data returned by APIs */
